code review, add 4 classes for work with OperationQueue
