+++ CloneVK/CloneVK/Helpers/OperationQueue/AsyncOperation.swift
@@ -1,0 +1,59 @@
+// AsyncOperation.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Асинхронная операция загрузки данных, вместо Operation,  с возможностью  управлять состоянием операции
+class AsyncOperation: Operation {
+    enum State: String {
+        case ready, executing, finished
+        fileprivate var keyPath: String {
+            "is" + rawValue.capitalized
+        }
+    }
+
+    // MARK: - Public Properties
+
+    var state = State.ready {
+        willSet {
+            willChangeValue(forKey: state.keyPath)
+            willChangeValue(forKey: newValue.keyPath)
+        }
+        didSet {
+            didChangeValue(forKey: state.keyPath)
+            didChangeValue(forKey: oldValue.keyPath)
+        }
+    }
+
+    // MARK: - Initializers
+
+    override var isAsynchronous: Bool {
+        true
+    }
+
+    override var isReady: Bool {
+        super.isReady && state == .ready
+    }
+
+    override var isExecuting: Bool {
+        state == .executing
+    }
+
+    override var isFinished: Bool {
+        state == .finished
+    }
+
+    override func start() {
+        if isCancelled {
+            state = .finished
+        } else {
+            main()
+            state = .executing
+        }
+    }
+
+    override func cancel() {
+        super.cancel()
+        state = .finished
+    }
+}
